requirements.txt
Pillow==10.1.0
numpy==1.26.2
pandas==2.1.4
albumentations==1.3.1
matplotlib==3.8.2
seaborn==0.11.2
tqdm==4.62.3
lion-pytorch==0.0.1
pandas==2.1.4
wandb==0.12.6
protobuf==3.20.0
lightning==2.1.2

--extra-index-url https://download.pytorch.org/whl/cu118
torch==2.1.0+cu118
--extra-index-url https://download.pytorch.org/whl/cu118
torchvision==0.16.0+cu118
